factoriel.rb Version Finale

+++ factoriel.rb
@@ -1,21 +1,14 @@
 # VARIABLES
 
-nbr = gets.chomp.to_i
+nbr = gets.chomp.to_f
 
 # FONCTIONS
 
-# def iterative_factoriel(n)
-#     puts (1..n).inject(:*)
-# end
-
 def recursive_factorial(n)
-    # Base case
     return 1 if n <= 1
-    # Recursive call
     n * recursive_factorial(n-1)
   end
 
 # CODE
 
-# iterative_factoriel(nbr)
 puts recursive_factorial(nbr)
